refactor([issue-1081]): split routes/pipeline.js god-router into domain sub-routers
server/routes/pipeline.js had grown to ~2,484 lines / 83 handlers. Split it
into server/routes/pipeline/{audio,series,arcs,manuscript,covers,issues,
editorial}.js assembled by index.js, mirroring the cos.js sub-router pattern.
Pure code motion: route paths, schemas, and handler bodies are unchanged
(verified by diffing the extracted method+path sets — identical 83 entries).
Shared error mapping (mapServiceError), the provider-override schema shape,
and countExtractedCanon moved to pipeline/shared.js. Comment pointers in
sharing.js, migrateSeriesCanon.js, tts.js, and promoteToPipeline.js updated
to the new file locations.

Refs #1081
